+++ packages/imperative/__tests__/config/cmd/auto-init/BaseAutoInitHandler.test.ts
@@ -19,6 +19,7 @@ import * as open from "open";
 import { ConfigSchema } from "../../../../../config";
 import { CredentialManagerFactory } from "../../../../../security";
 import { SessConstants } from "../../../../../rest";
+import { OverridesLoader } from "../../../../src/OverridesLoader";
 
 describe("BaseAutoInitHandler", () => {
     beforeEach( async () => {
@@ -71,7 +72,7 @@ describe("BaseAutoInitHandler", () => {
         };
         const mockSetSchema = jest.fn();
         const buildSchemaSpy = jest.spyOn(ConfigSchema, 'buildSchema').mockImplementation();
-        const ensureCredMgrSpy = jest.spyOn(handler as any, "ensureCredentialManagerLoaded");
+        const ensureCredMgrSpy = jest.spyOn(OverridesLoader, "ensureCredentialManagerLoaded");
         const displayAutoInitChangesSpy = jest.spyOn(handler as any, "displayAutoInitChanges");
         jest.spyOn(ImperativeConfig, 'instance', "get").mockReturnValue({
             config: {
@@ -148,7 +149,7 @@ describe("BaseAutoInitHandler", () => {
             }
         };
         const buildSchemaSpy = jest.spyOn(ConfigSchema, 'buildSchema').mockImplementation();
-        const ensureCredMgrSpy = jest.spyOn(handler as any, "ensureCredentialManagerLoaded");
+        const ensureCredMgrSpy = jest.spyOn(OverridesLoader, "ensureCredentialManagerLoaded");
         const displayAutoInitChangesSpy = jest.spyOn(handler as any, "displayAutoInitChanges");
         const mockSetSchema = jest.fn();
 
@@ -228,7 +229,7 @@ describe("BaseAutoInitHandler", () => {
             }
         };
         const buildSchemaSpy = jest.spyOn(ConfigSchema, 'buildSchema').mockImplementation();
-        const ensureCredMgrSpy = jest.spyOn(handler as any, "ensureCredentialManagerLoaded");
+        const ensureCredMgrSpy = jest.spyOn(OverridesLoader, "ensureCredentialManagerLoaded");
         const displayAutoInitChangesSpy = jest.spyOn(handler as any, "displayAutoInitChanges");
         const mockSetSchema = jest.fn();
 
@@ -308,7 +309,7 @@ describe("BaseAutoInitHandler", () => {
         });
         const mockSecureFields = jest.fn().mockReturnValue([]);
         const mockFindSecure = jest.fn().mockReturnValue([]);
-        const ensureCredMgrSpy = jest.spyOn(handler as any, "ensureCredentialManagerLoaded");
+        const ensureCredMgrSpy = jest.spyOn(OverridesLoader, "ensureCredentialManagerLoaded");
         const displayAutoInitChangesSpy = jest.spyOn(handler as any, "displayAutoInitChanges");
         const mockImperativeConfigApi = {
             layers: {
@@ -390,7 +391,7 @@ describe("BaseAutoInitHandler", () => {
             exists: true,
             properties: {}
         });
-        const ensureCredMgrSpy = jest.spyOn(handler as any, "ensureCredentialManagerLoaded");
+        const ensureCredMgrSpy = jest.spyOn(OverridesLoader, "ensureCredentialManagerLoaded");
         const displayAutoInitChangesSpy = jest.spyOn(handler as any, "displayAutoInitChanges");
         const mockImperativeConfigApi = {
             layers: {
@@ -466,7 +467,7 @@ describe("BaseAutoInitHandler", () => {
             exists: true,
             properties: {}
         });
-        const ensureCredMgrSpy = jest.spyOn(handler as any, "ensureCredentialManagerLoaded");
+        const ensureCredMgrSpy = jest.spyOn(OverridesLoader, "ensureCredentialManagerLoaded");
         const displayAutoInitChangesSpy = jest.spyOn(handler as any, "displayAutoInitChanges");
         const mockImperativeConfigApi = {
             layers: {
@@ -563,7 +564,7 @@ describe("BaseAutoInitHandler", () => {
         });
         const mockSecureFields = jest.fn().mockReturnValue(["profiles.base.properties.tokenValue"]);
         const mockFindSecure = jest.fn().mockReturnValue([]);
-        const ensureCredMgrSpy = jest.spyOn(handler as any, "ensureCredentialManagerLoaded");
+        const ensureCredMgrSpy = jest.spyOn(OverridesLoader, "ensureCredentialManagerLoaded");
         const displayAutoInitChangesSpy = jest.spyOn(handler as any, "displayAutoInitChanges");
         const mockImperativeConfigApi = {
             layers: {