+++ packages/imperative/src/OverridesLoader.ts
@@ -23,6 +23,8 @@ import { Logger } from "../../logger";
  * You should not need to call this from your CLI.
  */
 export class OverridesLoader {
+    private static readonly ZOWE_CLI_PACKAGE_NAME = "@zowe/cli";
+
     /**
      * Apply overrides to all applicable facilities and use our defaults where
      * an override is not provided.
@@ -35,37 +37,17 @@ export class OverridesLoader {
         packageJson: any
     ): Promise<void> {
         // Initialize the Credential Manager
-        await (ImperativeConfig.instance.config?.exists ? this.loadCredentialManager : this.loadCredentialManagerOld)(config, packageJson);
+        await this.loadCredentialManager(config, packageJson, ImperativeConfig.instance.config?.exists);
     }
 
     /**
-     * Load the baked-in zowe CredentialManager and initialize it.
-     * If we need to reinstate 3rd party overrides, delete this function and
-     * rename loadCredentialManagerOld.
-     *
-     * @internal
-     * @param {IImperativeConfig} config - the current {@link Imperative#loadedConfig}
-     * @param {any} packageJson - the current package.json
+     * Ensure the Credential Manager is initialized for team config.
      */
-    public static async loadCredentialManager(
-        config: IImperativeConfig,
-        packageJson: any
-    ): Promise<void> {
-        if (packageJson.dependencies?.keytar != null || packageJson.optionalDependencies?.keytar != null) {
-            // Load the CredentialManager built into Imperative for managing secure properties
-            await CredentialManagerFactory.initialize({
-                // The display name will be the plugin name that introduced the override OR it will default to the CLI name
-                displayName: config.productDisplayName || config.name,
-                // For overrides, the service could be the CLI name, but we do not override anymore.
-                // Null will use the service name of the built-in credential manager.
-                service: config.credentialServiceName || null,
-                // If the default is to be used, we won't implant the invalid credential manager.
-                // We do not use the invalid credential manager, since we no longer allow overrides.
-                invalidOnFailure: false
-            });
-        }
+    public static async ensureCredentialManagerLoaded(): Promise<void> {
+        if (CredentialManagerFactory.initialized) return;
 
-        await OverridesLoader.loadSecureConfig();
+        await this.loadCredentialManager(ImperativeConfig.instance.loadedConfig,
+            ImperativeConfig.instance.callerPackageJson, true);
     }
 
     /**
@@ -73,14 +55,14 @@ export class OverridesLoader {
      *
      * @param {IImperativeConfig} config - the current {@link Imperative#loadedConfig}
      * @param {any} packageJson - the current package.json
+     * @param {boolean} useTeamConfig - specify True if team config is active
      */
-    private static async loadCredentialManagerOld(
+    private static async loadCredentialManager(
         config: IImperativeConfig,
-        packageJson: any
+        packageJson: any,
+        useTeamConfig?: boolean
     ): Promise<void> {
         const overrides: IImperativeOverrides = config.overrides;
-
-        const ZOWE_CLI_PACKAGE_NAME = `@zowe/cli`;
 
         // The manager display name used to populate the "managed by" fields in profiles
         const displayName: string = (
@@ -97,7 +79,7 @@ export class OverridesLoader {
             config.productDisplayName || config.name;
 
         // Initialize the credential manager if an override was supplied and/or keytar was supplied in package.json
-        if (overrides.CredentialManager != null || packageJson.dependencies?.keytar != null) {
+        if (overrides.CredentialManager != null || this.shouldUseKeytar(packageJson, useTeamConfig)) {
             let Manager = overrides.CredentialManager;
             if (typeof overrides.CredentialManager === "string" && !isAbsolute(overrides.CredentialManager)) {
                 Manager = resolve(process.mainModule.filename, "../", overrides.CredentialManager);
@@ -108,16 +90,29 @@ export class OverridesLoader {
                 Manager,
                 // The display name will be the plugin name that introduced the override OR it will default to the CLI name
                 displayName,
-
-                // zowe cli will always add `Zowe` to it's list of service names
-                service: config?.name === ZOWE_CLI_PACKAGE_NAME ? DefaultCredentialManager.SVC_NAME : config?.name,
-
+                // The service is always the CLI name (Keytar and other plugins can use this to uniquely identify the service)
+                service: config.name === this.ZOWE_CLI_PACKAGE_NAME ? DefaultCredentialManager.SVC_NAME : config.name,
                 // If the default is to be used, we won't implant the invalid credential manager
                 invalidOnFailure: !(Manager == null)
             });
         }
 
         await OverridesLoader.loadSecureConfig();
+    }
+
+    /**
+     * Check if the DefaultCredentialManager which uses keytar should be enabled.
+     * We require that keytar is listed as a dependency in package.json, and one of the following is true:
+     *  1. AppSettings are not initialized (SDK usage)
+     *  2. Team config is active (CLI with v2 profiles)
+     *  3. CredentialManager override is host package name (CLI with v1 profiles)
+     * @param packageJson The current package.json of the CLI package
+     * @param useTeamConfig Specify True if team config is active
+     * @returns True if DefaultCredentialManager should be used
+     */
+    private static shouldUseKeytar(packageJson: any, useTeamConfig: boolean): boolean {
+        return (packageJson.dependencies?.keytar != null || packageJson.optionalDependencies?.keytar != null) &&
+            (!AppSettings.initialized || useTeamConfig || AppSettings.instance.getNamespace("overrides")?.CredentialManager === packageJson.name);
     }
 
     /**