+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to the Imperative package will be documented in this file.
 
 ## Recent Changes
 
-- **Breaking**: Use JSON-based communication protocol between imperative daemon server and client.
+- **Next Breaking**: Use JSON-based communication protocol between imperative daemon server and client.
+
+## `5.0.0-next.202111192150`
+
+- BugFix: Changed credentials to be stored securely by default for v1 profiles to be consistent with the experience for v2 profiles. [zowe/zowe-cli#1128](https://github.com/zowe/zowe-cli/issues/1128)
+- **Next Breaking**
+    - Removed the `credentialServiceName` property from ImperativeConfig. The default credential manager uses the `name` property instead.
 
 ## `5.0.0-next.202111101806`